feat: analytics ok
