+++ .env
@@ -4,5 +4,5 @@ DATABASE_USER=postgres
 DATABASE_PASSWORD=postgres
 DATABASE_NAME=postgres
 
-NODE_DOCKER_PORT=3000
+NODE_DOCKER_PORT=3001
