+++ .env.development
@@ -1,0 +1,10 @@
+DATABASE_HOST=analytics-db
+DATABASE_PORT=5432
+DATABASE_USER=postgres
+DATABASE_PASSWORD=postgres
+DATABASE_NAME=postgres
+
+NODE_DOCKER_PORT=3000
+
+URL=0.0.0.0
+PORT=5000
