+++ Dockerfile
@@ -3,5 +3,5 @@ WORKDIR /analytics-service
 COPY package.json ./
 RUN npm install
 COPY . ./
-CMD npm run start:debug
+CMD npm run start:dev
 EXPOSE 5000
