+++ docker-compose.yml
@@ -1,22 +1,38 @@
-version: '3.8'
+version: "3.8"
 services:
-  db:
-    image: postgres:12
-    environment:
-      - POSTGRES_HOST_AUTH_METHOD=trust
-    ports:
-      - 5432:5432
-    volumes:
-      - db:/var/lib/mysql
+    adminer:
+        image: adminer
+        restart: always
+        ports:
+            - 8080:8080
+        networks:
+            - fidecloud
+        volumes:
+            - db:/var/lib/postgresql/gui
 
-  web:
-    depends_on:
-      - db
-    build: ./
-    ports:
-      - 5000:5000
-    environment:
-      - .env
+    db:
+        image: postgres:12
+        environment:
+            - POSTGRES_HOST_ANALYTICS_METHOD=trust
+        ports:
+            - 5432:5432
+        volumes:
+            - db:/var/lib/postgresql/gui
+        networks:
+            - fidecloud
+
+    web:
+        depends_on:
+            - db
+        build: ./
+        ports:
+            - 5001:5001
+        environment:
+            - .env
 
 volumes:
-  db:
+    db:
+
+networks:
+    fidecloud:
+        driver: bridge
