+++ src/card/card.module.ts
@@ -1,6 +1,5 @@
-import { Module, forwardRef } from "@nestjs/common";
+import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { ShopModule } from "src/shop/shop.module";
 import { CardController } from "./card.controller";
 import { Card } from "./card.entity";
 import { CardService } from "./card.service";
@@ -8,6 +7,7 @@ import { CardService } from "./card.service";
 @Module({
   controllers: [CardController],
   providers: [CardService],
-  imports: [forwardRef(() => ShopModule), TypeOrmModule.forFeature([Card])],
+  imports: [TypeOrmModule.forFeature([Card])],
+  exports: [CardService, TypeOrmModule],
 })
 export class CardModule {}
